Low Resolution LHPD2a1 Neuron Meshes from BANC

A list of simplified mesh3d objects representing the 3D surface reconstructions of 4 LHPD2a1 neurons from the BANC connectome dataset.

Usage

banc.meshes

Format

A list of 5 mesh3d objects, each containing vertices and faces defining the 3D surface of a neuron.

Source

BANC connectome dataset, accessed via bancr::banc_read_neuron_meshes and decimated using Rvcg::vcgQEdecim.

References

Dolan, M. J., Belliart-Guérin, G., Bates, A. S., Frechter, S., Lampin-Saint-Amaux, A., Aso, Y., Roberts, R. J. V., Schlegel, P., Wong, A., Hammad, A., Bock, D., Rubin, G. M., Preat, T., Plaçais, P. Y., & Jefferis, G. S. X. E. (2018). Communication from learned to innate olfactory processing centers is required for memory retrieval in Drosophila. Neuron, 100(3), 651-668.e8. doi:10.1016/j.neuron.2018.08.037

Bates, A. S., Phelps, J. S., Kim, M., Yang, H. H., Matsliah, A., Ajabi, Z., Perlman, E., Dunne, J., Roat, J., Joyce, P., Bogovic, J. A., Jefferis, G. S. X. E., Murthy, M., Card, G., & The FlyWire Consortium. (2025). Distributed control circuits across a brain-and-cord connectome. bioRxiv. doi:10.1101/2025.07.31.667571

Examples

if (FALSE) { # \dontrun{
library(nat.ggplot)
library(ggplot2)

# Plot first neuron mesh
ggplot() +
  geom_neuron(banc.meshes[[1]],
              rotation_matrix = banc_view,
              cols = c("purple", "magenta"))

# Plot all neuron meshes
ggneuron(banc.meshes,
         rotation_matrix = banc_view,
         cols = c("purple", "magenta"))
} # }