Flow Centrality Split LHPD2a1 Neurons from BANC

A neuronlist containing LHPD2a1 neurons that have been split into axonal and dendritic compartments using flow centrality analysis. This method, described by Schneider-Mizell et al. (2016), uses the spatial distribution of pre- and postsynaptic sites to segregate neuronal arbors into functionally distinct compartments.

Usage

banc.neurons.flow

Format

A neuronlist object with neurons containing additional Label field in their data frames indicating compartment type:

Label = 2

Axon

Label = 3

Dendrite

Label = 4

Primary dendrite

Label = 7

Primary neurite

Label = 0 or NA

Unclassified

Source

BANC connectome dataset (Bates et al., 2025). Split using flow centrality method from hemibrainr package.

References

Schneider-Mizell, C. M., Gerhard, S., Longair, M., Kazimiers, T., Li, F., Zwart, M. F., Champion, A., Midgley, F. M., Fetter, R. D., Saalfeld, S., & Cardona, A. (2016). Quantitative neuroanatomy for connectomics in Drosophila. eLife, 5, e12059. doi:10.7554/eLife.12059

Bates, A. S., Phelps, J. S., Kim, M., Yang, H. H., Matsliah, A., Ajabi, Z., Perlman, E., Dunne, J., Roat, J., Joyce, P., Bogovic, J. A., Jefferis, G. S. X. E., Murthy, M., Card, G., & The FlyWire Consortium. (2025). Distributed control circuits across a brain-and-cord connectome. bioRxiv. doi:10.1101/2025.07.31.667571

Examples

if (FALSE) { # \dontrun{
library(nat.ggplot)
library(ggplot2)

# Plot split neurons showing axon/dendrite compartments
ggplot() + geom_neuron(banc.neurons.flow[[1]], rotation_matrix = banc_view)

# Plot all split neurons
ggneuron(banc.neurons.flow, rotation_matrix = banc_view)
} # }