LHPD2a1 Neuron Skeletons from BANC Connectome

A neuronlist containing skeletonised morphologies of 5 LHPD2a1 neurons from the BANC (Brain And Nerve Cord) connectome dataset. These neurons are lateral horn local neurons first described as LHPD2a1 by Dolan et al. (2018) in the adult fly brain. The skeletons have been processed using wavefront skeletonisation and rerooted.

Usage

banc.skels

Format

A neuronlist object containing 5 neurons with 3D point coordinates and connectivity information.

Source

BANC connectome dataset (Bates et al., 2025). Data accessed via the bancr package.

References

Dolan, M. J., Belliart-Guérin, G., Bates, A. S., Frechter, S., Lampin-Saint-Amaux, A., Aso, Y., Roberts, R. J. V., Schlegel, P., Wong, A., Hammad, A., Bock, D., Rubin, G. M., Preat, T., Plaçais, P. Y., & Jefferis, G. S. X. E. (2018). Communication from learned to innate olfactory processing centers is required for memory retrieval in Drosophila. Neuron, 100(3), 651-668.e8. doi:10.1016/j.neuron.2018.08.037

Bates, A. S., Phelps, J. S., Kim, M., Yang, H. H., Matsliah, A., Ajabi, Z., Perlman, E., Dunne, J., Roat, J., Joyce, P., Bogovic, J. A., Jefferis, G. S. X. E., Murthy, M., Card, G., & The FlyWire Consortium. (2025). Distributed control circuits across a brain-and-cord connectome. bioRxiv. doi:10.1101/2025.07.31.667571

Examples

if (FALSE) { # \dontrun{
library(nat)
library(nat.ggplot)

# Plot the neurons
plot3d(banc.skels)

# Use with ggplot2
library(ggplot2)
ggplot() + geom_neuron(banc.skels, rotation_matrix = banc_view)
} # }